qed.gql: group types into archives/readers/datasets/controllers subpackages
