+++ pkg/gql/archives/Item.py
@@ -8,7 +8,7 @@
 import graphene
 
 # my interface
-from .Node import Node
+from ..Node import Node
 
 
 # my node type
