+++ pkg/gql/controllers/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+#
+# michael a.g. aïvázis <michael.aivazis@para-sim.com>
+# (c) 1998-2026 all rights reserved
+
+
+# the controller graphql types are imported directly by their consumers to keep the
+# schema import graph acyclic; this package marker stays intentionally empty
+
+
+# end of file
