+++ pkg/gql/readers/ConnectReader.py
